requirements.txt
Django==5.2.8
psycopg2-binary==2.9.11
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.29.0
django-cors-headers==4.9.0
dj-database-url==3.0.1
whitenoise==6.11.0
gunicorn==23.0.0
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.3
